chore: refactor navbar component
