+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
-import Nav from './nav'
 
 const geistSans = Geist({
     variable: '--font-geist-sans',
