+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Nav from './nav'
+import Nav from '../components/nav'
 
 export default function Home() {
     return (
