+++ src/app/products/create/page.tsx
@@ -1,0 +1,20 @@
+import Nav from '@/components/nav'
+
+const CreateProduct = () => {
+    return (
+        <>
+            <Nav />
+            <h1>Create Product</h1>
+            {/**
+             * make
+             * model
+             * year
+             * price
+             * description
+             * milage
+             */}
+        </>
+    )
+}
+
+export default CreateProduct
