+++ src/app/products/page.tsx
@@ -1,0 +1,13 @@
+import Nav from '@/components/nav'
+
+const Products = () => {
+    return (
+        <>
+            <Nav />
+            <h1>Product List</h1>
+        </>
+        //TODO: button redirect to create of page
+    )
+}
+
+export default Products
