+++ src/components/nav.tsx
@@ -3,7 +3,9 @@ const Nav = () => {
         <nav className='flex justify-end py-5 px-20'>
             <ul className='flex space-x-5'>
                 <li>Home</li>
-                <li>Sell</li>
+                <li>
+                    <a href='/products/create'>Sell</a>
+                </li>
                 <li>About Us</li>
                 <li>Contact Us</li>
                 <li>Client Releases</li>
